benchmarks/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
bcrypt==3.1.7
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
colorful==0.5.0
cryptography==2.7
cycler==0.10.0
idna==2.8
isort==4.3.16
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
matplotlib==3.0.3
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
numpy==1.16.2
pandas==0.24.2
paramiko==2.6.0
pycparser==2.19
pylint==2.3.1
PyNaCl==1.3.0
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
requests==2.21.0
six==1.12.0
tqdm==4.31.1
typed-ast==1.3.1
urllib3==1.25.1
wrapt==1.11.1
yapf==0.28.0
